# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: stxact Proxy API
  version: 1.0.0
  description: |
    OpenAPI specification for the stxact proxy routes used by webapp and CLI.
servers:
  - url: http://localhost:3001
    description: Local development
...
paths:
...
  /receipts:
    get:
      tags: [Receipts]
      summary: List receipts
      parameters:
        - in: query
          name: seller_principal
          schema: { type: string }
        - in: query
          name: buyer_principal
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0, default: 0 }
        - in: query
          name: sort
          schema:
            type: string
            enum: [timestamp_desc, timestamp_asc]
      responses:
        "200":
          description: Receipt list
          content:
            application/json:
              schema:
                type: object
                properties:
                  receipts:
                    type: array
                    items:
                      $ref: "#/components/schemas/Receipt"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
  /receipts/{receipt_id}:
    get:
      tags: [Receipts]
      summary: Get receipt by ID
      parameters:
        - in: path
          name: receipt_id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Receipt payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Receipt"
        "404":
          description: Not found
  /receipts/verify:
    post:
      tags: [Receipts]
      summary: Verify receipt signature
      parameters:
        - in: query
          name: on_chain
          schema: { type: boolean }
        - in: query
          name: bns
          schema: { type: boolean }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                receipt:
                  $ref: "#/components/schemas/Receipt"
              required: [receipt]
      responses:
        "200":
          description: Verification results
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid: { type: boolean }
                  checks:
                    $ref: "#/components/schemas/VerificationChecks"
...
components:
  schemas:
    Pagination:
      type: object
      properties:
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
        has_more: { type: boolean }
...
    Receipt:
      type: object
      properties:
        receipt_id: { type: string, format: uuid }
        request_hash: { type: string }
        payment_txid: { type: string }
        seller_principal: { type: string }
        seller_bns_name: { type: string, nullable: true }
        buyer_principal: { type: string, nullable: true }
        delivery_commitment: { type: string, nullable: true }
        timestamp: { type: integer }
        block_height: { type: integer }
        block_hash: { type: string }
        key_version: { type: integer }
        revision: { type: integer }
        service_policy_hash: { type: string, nullable: true }
        signature: { type: string }
    VerificationChecks:
      type: object
      properties:
        signature_valid: { type: boolean }
        principal_match: { type: boolean, nullable: true }
        payment_txid_confirmed: { type: boolean, nullable: true }
        bns_verified: { type: boolean, nullable: true }
...